Allow using WEBFLOW_API_TOKEN env variable
If `WEBFLOW_API_TOKEN` is set, then the `--api-token` flag is no longer
required. It makes the use of the tool much easier as one can `export
WEBFLOW_API_TOKEN` in their bash session and just use the commands
without needing to repeatedly type the `--api-token` flag. If both the
`--api-token` and `WEBFLOW_API_TOKEN` are not set, the tool throws an
error.